Extended to support configurable HTTP endpoints
Extended the MetricsHandler class to support HTTP/1.1 and to allow the HTTP server endpoint to be specified during creation. With these changes, I am able to use the same URL to extract the metrics for all clients. I realize that the existing behavior would return the metrics for any URL at the specified port, but I want only a single URL for extraction and all others rejected.